server/requirements.txt
asgiref==3.11.1
celery==5.6.0
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
gunicorn==25.3.0
packaging==26.2
psycopg==3.3.3
psycopg-binary==3.3.3
redis==7.1.0
sqlparse==0.5.5
whitenoise==6.12.0
